+++ src/Polysemy.hs
@@ -3,6 +3,7 @@ module Polysemy
     Sem ()
   , Member
   , Members
+  , LastMember
 
   -- * Running Sem
   , run
@@ -88,6 +89,9 @@ module Polysemy
   , reinterpret2H
   , reinterpret3H
 
+    -- * Combinators for Interpreting Directly to IO
+  , withLowerToIO
+
     -- * Kind Synonyms
   , Effect
   , EffectRow
@@ -117,7 +121,9 @@ module Polysemy
 
 import Polysemy.Internal
 import Polysemy.Internal.Combinators
+import Polysemy.Internal.Forklift
 import Polysemy.Internal.Kind
 import Polysemy.Internal.TH.Effect
 import Polysemy.Internal.Tactics
+import Polysemy.Internal.Union