+++ src/Polysemy/Internal/Kind.hs
@@ -4,9 +4,13 @@ import Data.Kind
 
 ------------------------------------------------------------------------------
 -- | The kind of effects.
+--
+-- @since 0.5.0.0
 type Effect    = (Type -> Type) -> Type -> Type
 
 ------------------------------------------------------------------------------
 -- | The kind of effect rows.
+--
+-- @since 0.5.0.0
 type EffectRow = [Effect]